moving project list to configuration pkg
